No mex extension
Author: Tuomas Eerola

Recent PC version of the Matlab (7.4) seems to be imcompatible with the
C-function/mex file (read and write midi files) so versions of the
nmat2midi and midi2nmat that do not rely on mex-files is released here.
These files are simple and primitive and are partly based on code
developed by Ken Schutte (see http://www.kenschutte.com/software).
Further improvements are welcome.

https://www.jyu.fi/hum/laitokset/musiikki/en/research/coe/materials/miditoolbox/